Refactor files and target.
