+++ prodmodel/model/files/external_data_file.py
@@ -1,5 +1,5 @@
-from model.input_file import InputFile
-from model.external_data_target import ExternalDataTarget
+from model.files.input_file import InputFile
+from model.target.external_data_target import ExternalDataTarget
 
 from pathlib import Path
